requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.4
Django==5.1.2
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
h==0.99
idna==3.10
psycopg2==2.9.10
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
